use wait and notify
